John Dyson's patches (and a few from me too) to LFS to use a different buffering scheme and make it more in tune with FreeBSD's vfs_bio implementation.  The filesystem seems fairly stable, but I wouldn't recommend it to anyone not willing to experience problems.  This is very green code and has the limitation that YOU CAN ONLY HAVE ONE LFS PARTITION MOUNTED AT A TIME.
What LFS is good for:

	Non fsynced writes	FASTER THAN FFS
	Large deletions		Increadibly fast

Reads are a little bit slower than FFS right now, but that is a factor of
how under optimized this code is.  LFS should in theory perform at least as
well as FFS under fsync (iozone) type loads, and this is what I'm currently
working on.

Reviewed by:	Justin Gibbs
Submitted by:	John Dyson
Obtained from:


git-svn-id: svn+ssh://svn.freebsd.org/base/head@4535 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f