This is a purely cosmetic change but it outlines what I'm GOING to do to this file. It's sort of an unignorable notification of coming changes..
This is a bit rude I understand.. but I can't afford to haqve the
diskslice code  drifting off too much further from a workable system
and I think I need to jump in now to make it obvious what has to be done
before it's too late.

appologies to bruce in advance.


git-svn-id: svn+ssh://svn.freebsd.org/base/head@8194 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f