Reset ip_divert_ignore to zero immediately after use - also, set it in the first place, independent of whether sin->sin_port is set.
The result is that diverted packets that are being forwarded
will be diverted once and only once on the way in (ip_input())
and again, once and only once on the way out (ip_output()) -
twice in total.  ICMP packets that don't contain a port will
now also be diverted.