23:59's mega-merge for ppp(8).
I've merged the three changes:

    re-open on HUP (+doc)
    Working -background option
    pending sig changes

and enclosed patches to get from the current RELENG_2_2.
The pending signal stuff is the most important, without it it's possible
to corrupt the caused variable, resulting in a call to (*NULL)() :(

(I'm not totally comfortable with the SIGHUP changes, but they are not
very harmful either. -- Joerg)

Submitted by:	brian