- Ignore KBD_RAW_MODE and KBD_CODE_MODE in scgetc() when it's called from the low-level console routines sccngetc() and sccncheckc(). Submitted by: bde (a long time ago)
- Don't try to ring bell and immediately return from do_bell() while
device probe is in progress at boot time; the timeout queue is not
functional yet.
PR: kern/2424

- Stop running the screen saver after panic() is called: check
if `panicstr' is non-NULL during scrn_timer().
PR: kern/5314

- A new option: SC_DISABLE_REBOOT
The reboot key (usually Ctl-Alt-Del) will be ignored if this option
is defined. You may still have the reboot key defined in the keymap and
it won't cause error when the keymap is loaded, but it will be quietly
treated as nop.

OKed by: sos


git-svn-id: svn+ssh://svn.freebsd.org/base/head@33239 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f