Change the delivery mechanism for incoming target commands.  We now use a 256 entry ring buffer of descriptersfor this purpose.  This allows the use of a simple 8bit counter in the sequencer code for tracking start location.
Entries in the ring buffer now contain a "cmd_valid" byte at their tail.
As an entry is serviced, this byte is cleared by the kernel and set by
the sequencer during its dma of a new entry.  Since this byte is the last
portion of the command touched during a dma, the kernel can use this
byte to ensure the command it processes is completely valid.

The new command format requires a fixed sized DMA from the controller
to deliver which allowed for additional simplification of the sequencer
code.  The hack that required 1 SCB slot to be stolen for incoming
command delivery notification is also gone.


git-svn-id: svn+ssh://svn.freebsd.org/base/head@41299 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f