Support for Intel 450NX-based systems with more than one PCI bus (ie. most of them).
Many thanks to Kevin Van Maren for the work here, Intel for lending us
a 450NX system to work this out on, and several other folks for testing
the patches.  See the PR for an extensive discussion of the nature of
the problem and resolution.

PR:		kern/8928
Submitted by:	Kevin Van Maren <vanmaren@fast.cs.utah.edu>


git-svn-id: svn+ssh://svn.freebsd.org/base/head@41930 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f