The 4 gig version of the Seagate Medalist Pro seems to have the same trouble with tagged queueing as the 6.5 gig version.
So, I've added a quirk entry for it to limit it to two outstanding
transactions at a time, just like the 6.5G version.  While I'm at it, add a
quirk for the 9G version of the drive, since it most likely has the same
problem.

Submitted by:	Jeremy Lea <reg@shale.csir.co.za>


git-svn-id: svn+ssh://svn.freebsd.org/base/head@44567 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f