Don't discard error output from sysctl(8).
Do discard standard output from the sysctl for approxy_all, and echo
what this sysctl is doing in the usual way.  This fix is probably
backwards.  We should probably just use the standard sysctl output
in all cases (it needs to have a newline filtered out).

Echo what the sysctls for nfs_reserved_port_only and nfs_access_cache
are doing.


git-svn-id: svn+ssh://svn.freebsd.org/base/head@47755 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f