Add a hack to cam that makes the cam_xpt available to the rest of the kernel. Justin agress that there is no other reasonable alternative to do automatic rescans on connect.
The problem is that when a new device attaches to a SIM (SCSI host
controller) we need to send a XPT_SCAN_BUS command to the SIM using
xpt_action. This requires however that there is a peripheral available
to take the command (otherwise xpt_done and later bomb). The RESCAN
ioctl uses the same periph.

This enables a USB mass storage drive to do an automatic rescan on
connection of the drive.

The automatic dropping of a CAM entry on disconnection was already
working (asynchronous event).

The next thing to do is find someone to commit a change to vpo to do the
same thing. Just port umass_cam_rescan and friends across to that
driver.

Approved by:   gibbs


git-svn-id: svn+ssh://svn.freebsd.org/base/head@58972 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f